pyproject.toml
[project]
name = "codesitter"
version = "0.1.0"
description = "Real-time syntax-aware code indexing and search for TypeScript"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
homepage = "https://github.com/yourusername/codesitter"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "cocoindex>=0.1.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "aiofiles>=23.0.0",
    "watchdog>=3.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-language-pack>=0.1.0",
    "numpy>=1.24.0",
]

[project.scripts]
codesitter = "codesitter.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Use src layout
packages = ["src/codesitter"]

[tool.hatch.build.targets.sdist]
# Include additional files in source distribution
include = [
    "src/codesitter/**/*.py",
    "scripts/**/*.py",
    "docs/**/*.md",
    "README.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
